+++ src/components/Game/GameContent/GameContent.vue
@@ -1,8 +1,8 @@
 <template>
     <div id="game-content" class="d-flex flex-column">
         <GameContentHeader :game="game"/>
-        <GameContentPlayField :game="game" class="flex-grow-1"/>
-        <GameContentFooter :game="game"/>
+        <GameContentPlayField :game="game" :play="play" class="flex-grow-1" @playerVotes="playerVotes"/>
+        <GameContentFooter :game="game" :play="play"/>
     </div>
 </template>
 
@@ -19,6 +19,23 @@ export default {
         game: {
             type: Game,
             required: true,
+        },
+    },
+    data() {
+        return {
+            play: {
+                votes: [],
+            },
+        };
+    },
+    methods: {
+        playerVotes(vote) {
+            const idx = this.play.votes.findIndex(({ from }) => from === vote.from);
+            if (idx !== -1) {
+                return vote.for ? this.play.votes.splice(idx, 1, vote) : this.play.votes.splice(idx, 1);
+            } else if (vote.for) {
+                return this.play.votes.push(vote);
+            }
         },
     },
 };