+++ src/plugins/index.js
@@ -6,6 +6,8 @@ import "./v-tooltip";
 import "./vee-validate";
 import "./vue-flip";
 import i18n from "./vue-i18n";
+import "./vue-roller";
+import "./vue-select";
 import "./vue-toasted";
 
 export { Router, i18n };