+++ src/plugins/vue-roller/index.js
@@ -1,0 +1,4 @@
+import Vue from "vue";
+import VRoller from "vue-roller";
+
+Vue.component("VRoller", VRoller);