+++ src/plugins/vue-select/index.js
@@ -1,0 +1,5 @@
+import Vue from "vue";
+import VSelect from "vue-select";
+import "vue-select/src/scss/vue-select.scss";
+
+Vue.component("VSelect", VSelect);